Fix ansible-test python version usage.
